+++ README.md
@@ -22,17 +22,17 @@
    RUB / THB*  : X.XX
    ```
   данные записываются в память, для обновления требуется вручную обновить данные по котировкам
-* `/reset` - производит сброс внутренних значенией котировок, если их срок жизни превысил 1 час 1 мин (обновление курсов)
+* `/reset` - производит сброс внутренних значение котировок, если их срок жизни превысил 1 час 1 мин (обновление курсов)
 * `/usd` - производит запрос курса USD к RUB и записывает в память
 * `/thb` - производит запрос THB к USD и записывает в память
 * `/%` -  производит запрос соотношения котировок обмена RUB к THB. возможно задать свое значение для коммерческой помощи друзьям
-* `/ex` - запускает калькуулятор конверсии курса из RUB в THB и обратно
+* `/ex` - запускает калькулятор конверсии курса из RUB в THB и обратно
 
 
 ## Получение необходимых токенов
 
 Для корректной работы необходимо иметь несколько токенов для работы с API запросами / клиентами.
-Данные ключи возможно поместить в папку environments в файл  .env при запуске с машины, или в настройках ключей безопасности в Heroku
+Данные ключи возможно поместить в папку environments в файл  `.env` при запуске с машины, или в настройках ключей безопасности в Heroku
 
 #### Требуемые ключи:
 ```
@@ -43,26 +43,26 @@ URL_HEROKU     =  ХХХХХХХХ
 ```
 ### Детали получения ключей:
 
-<b><u>TELEBOT:</u></b>
+####TELEBOT:
 
 В телеграмме написать личное сообщение @BotFather - /newbot
 Пройти все последующие этапы регистрации бота.
 В конце бот выдаст сообщение с Токеном  и ссылками на API методы
   
-<b><u>TOKEN_BANGKOK:</b></u>
+####TOKEN_BANGKOK:
      
 ПО ссылке пройти регистрацию https://developer.bangkokbank.com/
 В разделе личного кабинета https://developer.bangkokbank.com/developer 
 найти токен для создания  API запросов Primary key
       
-<b><u>TOKEN_TINK:</b></u>
+####TOKEN_TINK:
      
 Перейдите в настройки профиля Тинькофф Инвестиции по ссылке: 
 https://www.tinkoff.ru/invest/settings/
       
-<b><u>URL_HEROKU:</b></u>
+####URL_HEROKU:
 
-соответсвует url вашего новосозданного приложения на Heroku
+Соответствует url вашего ново-созданного приложения на Heroku
 
    
 ## Heroku usage:
@@ -85,7 +85,7 @@ https://www.tinkoff.ru/invest/settings/
 * наслаждаться работой телебота
 
 ## Использование через компьютер 
-Запуск будет проивходить из файла``telebot_from_computer.py``
+Запуск будет происходить из файла``telebot_from_computer.py``
 
 Рекомендуется использовать для подстроек и проверки API ключей, или в случае развертывания на ином ресурсе от Heroku
 
@@ -101,14 +101,7 @@ https://www.tinkoff.ru/invest/settings/
 * произвести запуск `telebot_from_computer.py`
 * пройти в окно чата с telebot-ом и проверить работоспособность
 #### ВАЖНО
->при таком методе не нужно создавать `webhoock`, еcли до этого был запущен `webhook с Heroku` его нужно силой удалить.
+>при таком методе не нужно создавать `webhoock`,еcли до этого был запущен `webhook с Heroku` его нужно силой удалить.
 > URL_HEROKU создавать ненужно.
 
-###Тесты
-Присутствуют интегральные тесты для клинета Тинькофф банка.
-Для их запуска необходимо в файле `environments\.env` прописать корректный Токен для Тинькофф банка
-   ```
-   TOKEN_TINK     =  ХХХХХХХХ
-   ```
-Остальные тесты реализованы с использованием `mock` объектов
