+++ environments/.env
@@ -1,5 +1,5 @@
-TELEBOT=5479085555:AAHAClosZ4WJBOzHYCj2OHP56yoyRDiQQRA
-TOKEN_BANGKOK=9ba9b80ec8ac4a2a8e61a38f6db6058a
-TOKEN_TINK=t.To_-Jo9y_O_9DV5rnTgVh5uBb3mMSlsqoFR_COQsP6SEBxjeBf8lmSzfoh7Gw2ZlhB9EMhKUTzGtXKIJjX6sdQ
-URL_HEROKU=https://testbotherokuzdv.herokuapp.com
-BKKB_URL='https://bbl-sea-apim-p.azure-api.net/api/ExchangeRateService'
+TELEBOT=TELEBOT
+TOKEN_BANGKOK=TOKEN_BANGKOK
+TOKEN_TINK=TOKEN_TINK
+URL_HEROKU=URL_HEROKU
+BKKB_URL=BKKB_URL
