+++ tests/test_clients/test_bkkb_client.py
@@ -2,15 +2,15 @@ import pytest
 import datetime
 
 from pandas import DataFrame
-from tests.fixtures_for_bkkb_client import correct_client, incorrect_client, get_last_data_response,get_family_responce,get_lak_update_rates,get_usd_update_rates
 
 
 def test_token_bkkb(requests_mock, correct_client):
     bkkb_client = correct_client
     headers = bkkb_client.headers
     requests_mock.get('http://some_url/ServiceVersion', request_headers=headers, text='2.0.0.1', status_code=200)
-    assert bkkb_client.get_data('ServiceVersion').text == '2.0.0.1'
-    assert bkkb_client.get_data('ServiceVersion').status_code == 200
+    response = bkkb_client.get_data('ServiceVersion')
+    assert response.text == '2.0.0.1'
+    assert response.status_code == 200
 
 
 def test_incorrect_token_bkkb(requests_mock, incorrect_client):
@@ -18,8 +18,9 @@ def test_incorrect_token_bkkb(requests_mock, incorrect_client):
     headers = bkkb_client_incorrect.headers
     requests_mock.get('http://some_url/ServiceVersion', request_headers=headers, json={"message": None},
                       status_code=401)
-    assert bkkb_client_incorrect.get_data('ServiceVersion').json().get('message') is None
-    assert bkkb_client_incorrect.get_data('ServiceVersion').status_code == 401
+    response = bkkb_client_incorrect.get_data('ServiceVersion')
+    assert response.json().get('message') is None
+    assert response.status_code == 401
 
 
 def test_last_update(requests_mock, get_last_data_response, correct_client):
@@ -29,15 +30,16 @@ def test_last_update(requests_mock, get_last_data_response, correct_client):
     requests_mock.get('http://some_url/GetDateTimeLastUpdate', request_headers=headers, json=json_request,
                       status_code=200)
     now = datetime.datetime.now().strftime('%d/%m/%Y')
-    assert bkkb_client.get_last_update().json()[0].get('Day')[2:] == now[2:]
-    assert bkkb_client.get_last_update().status_code == 200
+    response = bkkb_client.get_last_update()
+    assert response.json()[0].get('Day')[2:] == now[2:]
+    assert response.status_code == 200
 
 
-def test_get_all_families(correct_client, requests_mock, get_family_responce):
+def test_get_all_families(correct_client, requests_mock, get_family_response):
     bkkb_client = correct_client
     headers = bkkb_client.headers
-    json_responce = get_family_responce
-    requests_mock.get('http://some_url/Getfxfamily', request_headers=headers, json=json_responce, status_code=200)
+    json_response = get_family_response
+    requests_mock.get('http://some_url/Getfxfamily', request_headers=headers, json=json_response, status_code=200)
     client_response = bkkb_client.get_all_values_families()
     assert client_response.status_code == 200
     assert len(client_response.json()) > 10
@@ -58,7 +60,7 @@ def test_get_all_families(correct_client, requests_mock, get_family_responce):
 def test_last_rate_update(correct_client, requests_mock, get_lak_update_rates):
     bkkb_client = correct_client
     headers = bkkb_client.headers
-    json_responce = get_lak_update_rates
+    json_response = get_lak_update_rates
     now = datetime.datetime.now()
     tdd = now.day
     tmm = now.month
@@ -69,7 +71,7 @@ def test_last_rate_update(correct_client, requests_mock, get_lak_update_rates):
     lang = 'en'
     family = 'LAK'
     requests_mock.get(f'http://some_url/GetChartfxrates/{fdd}/{fmm}/{fyyyy}/{tdd}/{tmm}/{tyyyy}/{family}/{lang}',
-                      request_headers=headers, json=json_responce, status_code=200)
+                      request_headers=headers, json=json_response, status_code=200)
     response = bkkb_client.get_last_all_rate_update_for_value(dict(day=fdd, month=fmm, year=fyyyy), family)
     assert response.status_code == 200
     assert len(response.json()) > 0
@@ -79,7 +81,7 @@ def test_last_rate_update(correct_client, requests_mock, get_lak_update_rates):
 def test_last_rate_update_incorrect_value(correct_client, requests_mock):
     bkkb_client = correct_client
     headers = bkkb_client.headers
-    json_responce = []
+    json_response = []
     now = datetime.datetime.now()
     tdd = now.day
     tmm = now.month
@@ -90,7 +92,7 @@ def test_last_rate_update_incorrect_value(correct_client, requests_mock):
     lang = 'en'
     family = 'INCORRECT_VALUE'
     requests_mock.get(f'http://some_url/GetChartfxrates/{fdd}/{fmm}/{fyyyy}/{tdd}/{tmm}/{tyyyy}/{family}/{lang}',
-                      request_headers=headers, json=json_responce, status_code=200)
+                      request_headers=headers, json=json_response, status_code=200)
     response = bkkb_client.get_last_all_rate_update_for_value(dict(day=fdd, month=fmm, year=fyyyy), family)
     assert response.status_code == 200
     assert len(response.json()) == 0
